refactor: move api to separate folder
